Refactor Era encoding/decoding
Move the EraReader/EraWriter from the "scale" pkg to "scaletypes"
