+++ polkaj-scale-types/src/main/java/io/emeraldpay/polkaj/scaletypes/EraReader.java
@@ -1,4 +1,4 @@
-package io.emeraldpay.polkaj.scale.reader;
+package io.emeraldpay.polkaj.scaletypes;
 
 import io.emeraldpay.polkaj.scale.ScaleCodecReader;
 import io.emeraldpay.polkaj.scale.ScaleReader;
