+++ polkaj-scale-types/src/main/java/io/emeraldpay/polkaj/scaletypes/EraWriter.java
@@ -1,4 +1,4 @@
-package io.emeraldpay.polkaj.scale.writer;
+package io.emeraldpay.polkaj.scaletypes;
 
 import io.emeraldpay.polkaj.scale.ScaleWriter;
 import io.emeraldpay.polkaj.scale.ScaleCodecWriter;
